+++ bin/seed_edx_database.py
@@ -1,7 +1,5 @@
 """Seed the edx database with test data."""
 
-import asyncio
-
 from sqlalchemy import create_engine
 
 from mongoengine import connect, disconnect
@@ -13,31 +11,30 @@ from mork.edx.mysql.factories.base import BaseSQLAlchemyModelFactory, Session
 from mork.edx.mysql.models.base import Base
 
 
-async def seed_edx_mysql_database():
+def seed_edx_mysql_database(batch_size: int = 100):
     """Seed the MySQL edx database with mocked data."""
     engine = create_engine(settings.EDX_MYSQL_DB_URL)
     Session.configure(bind=engine)
     BaseSQLAlchemyModelFactory._meta.sqlalchemy_session = Session  # noqa: SLF001
     Base.metadata.create_all(engine)
 
-    EdxAuthUserFactory.create_batch(100)
+    users = EdxAuthUserFactory.create_batch(batch_size)
+    usernames = [user.username for user in users]
+
     Session.commit()
+    return usernames
 
 
-async def seed_edx_mongodb_database():
-    """Seed the MongoDB edx database with mocked data."""
-    connect(host=settings.EDX_MONGO_DB_HOST, db=settings.EDX_MONGO_DB_NAME)
+def seed_edx_mongodb_database(batch_size: int = 1000, usernames: list[str] = []):
+    """Seed the MongoDB edx database with mocked data using existing usernames."""
+    connect(host=settings.EDX_MONGO_DB_URL)
 
-    CommentFactory.create_batch(1000)
-    CommentThreadFactory.create_batch(1000)
+    CommentFactory.create_batch(batch_size, usernames=usernames)
+    CommentThreadFactory.create_batch(batch_size, usernames=usernames)
 
-    disconnect(alias="mongodb")
-
-
-async def main():
-    tasks = [seed_edx_mysql_database(), seed_edx_mongodb_database()]
-    await asyncio.gather(*tasks)
+    disconnect()
 
 
 if __name__ == "__main__":
-    asyncio.run(main())
+    usernames = seed_edx_mysql_database()
+    seed_edx_mongodb_database(usernames=usernames)